KAFKA-12257; Consumer mishandles topics deleted and recreated with the same name (#11004)
Store topic ID info in consumer metadata. We will always take the topic ID from the latest metadata response and remove any topic IDs from the cache if the metadata response did not return a topic ID for the topic. The benefit of this is that it lets us detect topic recreations. This allows the client to update metadata even if the leader epoch is lower than what was seen previously.

Reviewers: Jason Gustafson <jason@confluent.io>