MINOR: Use parameterized logging in StandardAuthorizer and StandardAuthorizerData (#12192)
This updates StandardAuthorizer and StandardAuthorizerData to use parameterized logging per the SLF4J recommendation (see https://www.slf4j.org/faq.html). This also removes a couple if statements that explicitly check if trace is enabled, but the logger should handle not publishing the message and not constructing the String if trace is not enabled.

Reviewers: Jason Gustafson <jason@confluent.io>