MINOR: Use `LogConfig.validate` instead of `validateValues` in `KafkaMetadataLog` (#13051)
`LogConfig.validateValues` may fail or incorrectly succeed if the properties don't include defaults.

During the conversion of `LogConfig` to Java (#13049), it became clear that the `asInstanceOf[Long]`
calls in `LogConfig.validateValues` were converting `null` to `0` when this method was invoked
from `KafkaMetadataLog`. This means that it would be possible for it to validate successfully
in cases where it should not.

Reviewers: José Armando García Sancio <jsancio@apache.org>